requirements.txt
selenium==4.22.0
numpy==1.23.5
#tensorflow==2.16.1
xgboost==2.1.0
pandas==2.2.2
scikit-learn==1.5.0
google-cloud-storage==2.17.0
google-cloud-bigquery==3.25.0
db-dtypes==1.2.0
google-cloud-bigquery-storage==2.25.0
plotly==5.22.0
